requirements.txt
Flask==1.1.1
pandas==1.0.5
numpy==1.18.5
matplotlib==3.2.2
scikit-learn==0.22.2.post1
Flask-WTF==0.14.3
PyMySQL==0.10.0
google-cloud-storage
gcsfs
Flask-MonitoringDashboard
